+++ wazo_call_logd/database/queries/call_log.py
@@ -40,6 +40,7 @@ class ListParams(TypedDict, total=False):
     user_uuids: list[str]
     terminal_user_uuids: list[str]
     recorded: bool
+    requested_internal_extension: str
 
 
 class CallLogDAO(BaseDAO):
@@ -125,6 +126,14 @@ class CallLogDAO(BaseDAO):
             )
         else:
             query = session.query(CallLog)
+        if requested_internal_exten := params.get('requested_internal_exten'):
+            query = query.filter(
+                CallLog.requested_internal_exten == requested_internal_exten
+            )
+        if requested_internal_context := params.get('requested_internal_context'):
+            query = query.filter(
+                CallLog.requested_internal_context == requested_internal_context
+            )
 
         query = query.options(
             joinedload('participants'),