+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 25.12
+
+* CDR can be filtered using the following query string arguments:
+  * `requested_internal_extension`
+  * `requested_internal_context`
+
 ## 24.13
 
 * The CDR resource now contains a new field called `requested_user_uuid`